Minimal change to support for CHAR and VARCHAR types
This is a quick fix to support the above two types added in Impala 2.0.
The thrift definitions in the Impala 2.0 branch have too much to use the
script thrift/process_thrift.sh without making much larger changes. So
instead of using the process script, I manually editied the
cli_service.thrift to include the new types. This is much easier but
means the next time process_thrift is used, these changes will be lost.