Add support for named, qmark and numeric parameters in cursor.execute()
The paramstyle of the dbapi implementation is still pyformat, but now it support named, qmark and numeric parameters as well.
If the parameter argument to execute() is a list or tuple, numeric or qmark style is assumed. If it is dict it assumes pyformat or named parameters. It raises a ProgrammingError for any other type.
In the case of qmark, the replacement function now raises a ProgrammingException if the number of question marks in the operation does not match the size of the argument list or tuple.