Add a changelog of recent commits before cutting the next release.
It doesn't appear that impyla has ever included a change log in the manifest.
Since there have been so many changes, and also many requests about which
issues are getting fixed, it seems like a good time to amend that.

Based on some discussions I've had with a few on this projects, it also
feels like it might be a good time to bump the minor version to 15. In order
to avoid the types of problems we had before with regard to releasing
0.14.2, I'm also proposing that we adopt a pre-release version string at
the outset, although I'm not sure whether an alpha release (i.e., 0.15a1)
or an "rc" release (i.e., 0.15rc1) makes the most sense. In either case,
the package should only install if a user specifically requests the version.
Bare "pip install impyla" should still grab the current official release.