Ugprade Thift to 0.11.0 for Python 2
All files in impala/_thrift_gen were generated with Impala's
Thrift 0.11.0 compiler from the .thrift files in the Impyla repo.

It is possible now to use the same Thrift lib from Python 3 (Python 3
is supported from Thrift 0.10.0). This allows removing dependency on
thriftpy2 and a lot of code to handle the differences, but I want to
keep the two commits separate.

Testing:
- ran py.test --connect impala with a current Impala and one where
  Thrift was also upgraded to 0.11.0