Add no_utf8strings to thrift compiler option
Switching to thrift 0.11.0 caused calling .decode('utf-8','replace')
in Python 2.* on every deserialized string value, which meant that
Impyla returned 'unicode' instead of 'string' type for strings in
result sets. Thrift does this to keep Python 2 and 3 behavior as
similar as possible, but in our case this is a potential breaking
change.

The fix is to set no_utf8strings option in the thrift compiler.
An alternative would have been to re encode 'unicode' results to
'string' before returning the results, but that could lead to
regression in performance and break usages of strings that are
not valid utf8.

Other changes:
- a unicode test is added to test_data_types.py which would fail
  in Python 2.* without this fix
- process_thrift.sh is changed to not sync thrift files from Impala
  by default, so now it is a simple tool to recompile thrift files