Add support for authentication via JWT token
Impala recently added support for authentication via a JWT
token in IMPALA-10489. This allows a user to authenticate
with a separate trusted authentication service, obtain
a JWT, and pass that JWT to Impala on the Authencation
Bearer HTTP header.

This implements support for connecting with auth_mechanism=JWT,
which requires a 'jwt' argument specifying the standard string
encoding for the JWT. This is passed on the Authentication Bearer
HTTP header on each message.

Fixes #462

Testing:
 - Added unit tests in test_dbapi_connect.py
 - Tested against a running Impala cluster with JWT authentication