Update ImpalaService.thrift
ImpalaService.thrift is updated to contain CloseImpalaOperation,
which can be used get the number of modified rows in DMLs.
This is not just a  copy, some parts of ImpalaService.thrift are
not included to avoid pulling in more Thrift files as dependencies.

Also updated process_thrift.sh to work with current Impala env vars.