Avoid retrying non-idempotent RPCs in binary connections (#549)
See the #549 for the detailed analyses of the issue.

The fix works similarly to the existing solution for http connections:
- each RPC knows whether it is idempotent
- if the error comes from establishing the connection, then retry
- if the error comes from executing the RPC, only retry if the RPC
  is idempotent

A test is added that relies on slow metadata handling in the
Impala cluster to trigger timouts. It would be nice to add wider
and more reliable tests in the future similarly to the http tests
in test_hs2_fault_injection.py