Subtract RPC time from sleep in _wait_to_finish
The goal is to support "long poll" (IMPALA-13294). When query option
long_polling_time_ms is set, the impala server will wait in
GetOperationStatus for this time (or until the query status changes).
This allows detecting earlier that a query has finished without making
GetOperationStatus RPCs more frequent.

If long_polling_time_ms is not used then the effect should be minor -
GetOperationStatus is quick RPC so the time it takes should mainly
come from network delay.

_get_sleep_interval() is not changed (min 0.01s, max 1s) to avoid
regression in existing use cases. It could be useful to override
this in a later patch based on the value of long_polling_time_ms.