Fallback to winkerberos if kerberos is not available (#587)
Winkerberos is an alternative to kerberos package on Windows and
has the same api, so can be used as a drop in replacement.

Based on another PR:
https://github.com/cloudera/impyla/pull/504

The difference is that the current patch prefers kerberos if both are
available to avoid the breaking existing workflows. setup.py is
also not modified. To use winkerberos it has to be installed
independently from impyla and impyla should be installed without
[kerberos] extra.